# Optimizations and cleanups