Optimizes some hot spots seen during scale testing. The largest changes are around optimizing `toProto`.

```
name                           old time/op    new time/op    delta
MultiToProto/cellblocks-16       7.44µs ± 1%    5.40µs ± 1%  -27.34%  (p=0.000 n=9+9)
MultiToProto/no_cellblocks-16    5.22µs ± 0%    5.22µs ± 1%     ~     (p=0.423 n=10+10)

name                           old alloc/op   new alloc/op   delta
MultiToProto/cellblocks-16       6.63kB ± 0%    3.42kB ± 0%  -48.40%  (p=0.000 n=10+10)
MultiToProto/no_cellblocks-16    3.84kB ± 0%    3.84kB ± 0%     ~     (all equal)

name                           old allocs/op  new allocs/op  delta
MultiToProto/cellblocks-16         70.0 ± 0%      58.0 ± 0%  -17.14%  (p=0.000 n=10+10)
MultiToProto/no_cellblocks-16      64.0 ± 0%      64.0 ± 0%     ~     (all equal)
```