@dethi 2022-08-16
You will need to rebase after https://github.com/tsuna/gohbase/pull/195 is merged, that will fix the lint errors.