@dethi 2022-08-16
Eh the data race is because of the embedded context again https://github.com/tsuna/gohbase/issues/174 😞