@dethi 2022-08-16
If you call `m.SetContext(nil)` in `freeMulti` that would unblock here. I don't think you should try to fix https://github.com/tsuna/gohbase/issues/174 in this PR. It's a much bigger change (a welcome change but probably best to have it separated)