@aaronbee 2022-08-17
Oh, wow, I see the issue now. That is a bit scary. I'll leave out zeroing out the ctx field in freeMulti until #174 is resolved.